Controller/Controller_Service/requirements.txt
django>=2.0
django-extensions>=2.0
psycopg2-binary
pyjwt
django-rest-framework
pika
requests
django-cors-headers
python-daemon
